move actionfurtherevidence directory to furtherevidence one
to keep all related FE stuff organised in one directory
