+++ src/test/java/uk/gov/hmcts/reform/sscs/service/conversion/WordDocumentConverterTest.java
@@ -6,6 +6,7 @@ import static org.junit.Assert.assertNotEquals;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
+import java.util.Objects;
 import okhttp3.*;
 import org.apache.commons.io.FilenameUtils;
 import org.apache.pdfbox.io.IOUtils;
@@ -37,7 +38,8 @@ public class WordDocumentConverterTest {
         InputStream file = ClassLoader.getSystemResourceAsStream("wordDocument.doc");
 
         return new Response.Builder()
-                .body(ResponseBody.create(MediaType.get("application/pdf"), IOUtils.toByteArray(file)))
+                .body(ResponseBody.create(IOUtils.toByteArray(Objects.requireNonNull(file)),
+                    MediaType.get("application/pdf")))
                 .request(chain.request())
                 .message("")
                 .code(200)
